requirements.txt
arrow==1.2.0
asgiref==3.4.1
blessed==1.19.0
Django==3.2.7
django-picklefield==3.0.1
django-q==1.3.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
psycopg2-binary==2.9.1
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
wcwidth==0.2.5
